moved Configuration into the Auth Module removed package-info.java
