+++ Android/app/src/main/java/com/dk4max/HS_Esslingen/communitygroups/Auth/Configuration.java
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package com.dk4max.HS_Esslingen.communitygroups;
+package com.dk4max.HS_Esslingen.communitygroups.Auth;
 
 import android.content.Context;
 import android.content.Intent;
@@ -24,6 +24,7 @@ import android.text.TextUtils;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import com.dk4max.HS_Esslingen.communitygroups.ConnectionBuilderForTesting;
 import com.dk4max.HS_Esslingen.communitygroups.R;
 
 import net.openid.appauth.connectivity.ConnectionBuilder;
